+++ XXXX-service/src/main/java/com/framework/service/service/system/impl/SystemButtonServiceImpl.java
@@ -382,7 +382,7 @@ public class SystemButtonServiceImpl extends BaseService implements SystemButton
      */
     @Override
     public SystemButton getByIdParam(Long id) {
-        if (id == null || id < NumeralUtil.MULTIPLEXING_LONG_POSITIVE_ONE) {
+        if (id == null || id.longValue() < NumeralUtil.MULTIPLEXING_LONG_POSITIVE_ONE) {
             return null;
         }
         return this.selectByPrimaryKey(id);
