+++ XXXX-service/src/main/java/com/framework/service/service/system/impl/SystemLogServiceImpl.java
@@ -144,7 +144,7 @@ public class SystemLogServiceImpl extends BaseService implements SystemLogServic
      */
     @Override
     public SystemLog getByIdParam(Long id) {
-        if (id == null || id < NumeralUtil.MULTIPLEXING_LONG_POSITIVE_ONE) {
+        if (id == null || id.longValue() < NumeralUtil.MULTIPLEXING_LONG_POSITIVE_ONE) {
             return null;
         }
         return this.selectByPrimaryKey(id);
